Move the files from aggregator-processor subfolder to aggregator folder
