+++ aggregator/indexing_aggregator_processor.go
@@ -1,4 +1,4 @@
-package aggregatorprocessor
+package aggregator
 
 import (
 	"sync"
