+++ examples/incredible-dot-product/aggregator/main.go
@@ -6,7 +6,6 @@ import (
 	"os"
 
 	"github.com/Layr-Labs/eigensdk-go/aggregator"
-	aggregatorprocessor "github.com/Layr-Labs/eigensdk-go/aggregator/aggregator-processor"
 	"github.com/Layr-Labs/eigensdk-go/chainio/txmgr"
 	"github.com/Layr-Labs/eigensdk-go/logging"
 	taskmanager "github.com/Layr-Labs/eigensdk-go/task-manager"
@@ -104,8 +103,8 @@ func main() {
 	}
 
 	// iv. Create the Aggregator Processor. Here we use the IndexingAggregatorProcessor, you can see its implementation
-	// in aggregator/aggregator-processor/indexing_aggregator_processor.go
-	aggregatorProcessor, err := aggregatorprocessor.NewIndexingAggregatorProcessor(logger, taskResponder)
+	// in aggregator/indexing_processor.go
+	aggregatorProcessor, err := aggregator.NewIndexingAggregatorProcessor(logger, taskResponder)
 	if err != nil {
 		logger.Errorf("Failed to create Aggregator Processor: %w", err)
 		return
